requirements.txt
Flask==2.2.5
scikit-learn==1.3.2
pandas==2.2.2
tldextract==3.4.0
python-whois==0.7.3
requests==2.31.0
wtforms==3.0.1
email-validator==2.0.0
cryptography==41.0.3